+++ docs/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [v0.2.3.3] 
+- Updated `sftp_tools.download_file` function
+
+## [v0.2.3.2]
+- Fixed relative importing error
+
 ## [v0.2.3.1]
 - Fixed misspelling relative import of catalog